Allow provide core docker image and fluent-bit image.
Extend create core_instance in k8s provider to allow
setting up the specific core_instance docker image that
has to be used when deploying core itself.

Also implements the client side of cloud v1.2.3
that allows setting up an image for the core_instance
that will be used as the default image for all the pipelines
that belongs to this instance.

Signed-off-by: Jorge Niedbalski <j@calyptia.com>